argonvale-backend/pyproject.toml
[project]
name = "argonvale-backend"
version = "0.1.0"
description = "Backend for Argonvale game using FastAPI and PSPF"
authors = [
    {name = "Joseph Hall", email = "contact@josephmatthewhall.com"},
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pspf", 
    "python-multipart>=0.0.9",
    "websockets>=12.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
